pyproject.toml
[project]
name = "therapist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dispatch>=0.2.2",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.0",
    "websocket-client>=1.8.0",
]
